Re-enable reading .osm.gz from stdin.
Note that there is no --cache parameter any more, so that this is only useful for small
files and you must supply the --split-file argument.

The PBF format is the new cache format, just convert to that instead. Splitter cannot currently
do that but will be able to one day.

git-svn-id: http://svn.mkgmap.org.uk/splitter/trunk@165 838fabe5-b1a8-4b81-9896-7a84386980e9