splitter-element-version-v1.patch : adds the option --handle-element-version with three possible values "keep", "fake", and "remove". With "remove" (the default) splitter works like r428, means, the version of OSM elements is not written to the output files, The change should not have an impact regarding performance.
With "keep" splitter writes the version that is found in the input file.
With "fake" splitter sets the version to 1.

Open questions: what to do when the input file(s) contain elements
without (or invalid) version info and option keep is used.
Another special case is the pbf format:
Since r428 splitter always writes version 1 for nodes, this was a work-around
for a problem in the JOSM pbf plugin. 
This is still done even with --handle-element-version=remove

A last special case: with output=xml and
--handle-element-version=keep  or fake splitter will write api version 0.6 instead of 0.5


git-svn-id: http://svn.mkgmap.org.uk/splitter/trunk@430 838fabe5-b1a8-4b81-9896-7a84386980e9