- remove SparseLong2ShortMapHuge, - let SparseLong2ShortMapInline and SparseLong2IntMap use nearly the same code
The code in SparseLong2ShortMapHuge was introduced to fix a problem with large input files which write a huge number of nodes into the same area. This happens e.g.
with SRTM data or when you split planet with --num-tiles=2.
The new code uses a better trick to circumvent this problem and thus has a smaller memory footprint.

git-svn-id: http://svn.mkgmap.org.uk/splitter/branches/refactoring2@447 838fabe5-b1a8-4b81-9896-7a84386980e9